renamed page-viewer component name
